+++ src/models/ledger/objects/mod.rs
@@ -3,7 +3,9 @@ pub mod amendments;
 pub mod amm;
 pub mod bridge;
 pub mod check;
+pub mod credential;
 pub mod deposit_preauth;
+pub mod did;
 pub mod directory_node;
 pub mod escrow;
 pub mod fee_settings;
@@ -28,8 +30,10 @@ use amendments::Amendments;
 use amm::AMM;
 use bridge::Bridge;
 use check::Check;
+use credential::Credential;
 use deposit_preauth::DepositPreauth;
 use derive_new::new;
+use did::DID;
 use directory_node::DirectoryNode;
 use escrow::Escrow;
 use fee_settings::FeeSettings;
@@ -65,6 +69,8 @@ pub enum LedgerEntryType {
     AMM = 0x0079,
     Bridge = 0x0069,
     Check = 0x0043,
+    DID = 0x0049,
+    Credential = 0x0081,
     DepositPreauth = 0x0070,
     DirectoryNode = 0x0064,
     Escrow = 0x0075,
@@ -93,6 +99,8 @@ pub enum LedgerEntry<'a> {
     AMM(AMM<'a>),
     Bridge(Bridge<'a>),
     Check(Check<'a>),
+    DID(DID<'a>),
+    Credential(Credential<'a>),
     DepositPreauth(DepositPreauth<'a>),
     DirectoryNode(DirectoryNode<'a>),
     Escrow(Escrow<'a>),