+++ src/models/requests/account_objects.rs
@@ -14,6 +14,10 @@ use super::{CommonFields, LedgerIndex, LookupByLedgerRequest, Marker, Request};
 #[serde(rename_all = "snake_case")]
 pub enum AccountObjectType {
     Check,
+    #[serde(rename = "did")]
+    #[strum(serialize = "did")]
+    DID,
+    Credential,
     DepositPreauth,
     Escrow,
     Offer,