+++ src/models/requests/mod.rs
@@ -40,8 +40,7 @@ pub mod tx;
 pub mod unsubscribe;
 pub mod vault_info;
 
-use alloc::boxed::Box;
-use alloc::{borrow::Cow, string::String};
+use alloc::{borrow::Cow, boxed::Box, string::String};
 use derive_new::new;
 use serde::{Deserialize, Serialize};
 use serde_with::skip_serializing_none;